Update 7. Agriculture.ino

+++ examples/NodeMCU-32S/Projects/7. Agriculture/7. Agriculture.ino
@@ -80,7 +80,7 @@ void loop() {
 
     //Publish telemetry data 1
     float h = dht.readHumidity();
-    float t = dht.readTemperature();
+    int t = dht.readTemperature();
 
     JSONVar payloadObject;
     payloadObject["Humidity"] = h;
